requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
Pillow==7.2.0
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.4.3
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
